LPS-146089 Move Normalizer code to new module
