+++ modules/apps/normalizer/normalizer-impl/src/main/java/com/liferay/normalizer/internal/NormalizerImpl.java
@@ -12,30 +12,31 @@
  * details.
  */
 
-package com.liferay.util;
+package com.liferay.normalizer.internal;
 
 import com.liferay.ibm.icu.text.Transliterator;
+import com.liferay.normalizer.Normalizer;
 import com.liferay.portal.kernel.util.StringUtil;
 
 /**
  * @author Brian Wing Shun Chan
  * @author Shuyang Zhou
- * @see    com.liferay.rss.util.Normalizer
  */
-public class Normalizer {
+public class NormalizerImpl implements Normalizer {
 
-	public static String normalizeToAscii(String s) {
+	@Override
+	public String normalizeToAscii(String s) {
 		if (!_hasNonasciiCode(s)) {
 			return s;
 		}
 
-		String normalizedText = TransliteratorHolder.transform(s);
+		String normalizedText = _transliterator.transform(s);
 
 		return StringUtil.replace(
 			normalizedText, _UNICODE_TEXT, _NORMALIZED_TEXT);
 	}
 
-	private static boolean _hasNonasciiCode(String s) {
+	private boolean _hasNonasciiCode(String s) {
 		for (int i = 0; i < s.length(); i++) {
 			if (s.charAt(i) > 127) {
 				return true;
@@ -45,21 +46,13 @@ public class Normalizer {
 		return false;
 	}
 
-	private static final char[] _NORMALIZED_TEXT = {'l', '\'', '\"'};
+	private static final String[] _NORMALIZED_TEXT = {"l", "'", "\""};
 
-	private static final char[] _UNICODE_TEXT = {'\u0142', '\u02B9', '\u02BA'};
+	private static final String[] _UNICODE_TEXT = {
+		"\u0142", "\u02B9", "\u02BA"
+	};
 
-	private static class TransliteratorHolder {
-
-		public static String transform(String s) {
-			return _transliterator.transform(s);
-		}
-
-		private static final Transliterator _transliterator =
-			Transliterator.getInstance(
-				"Greek-Latin; Cyrillic-Latin; NFD; [:Nonspacing Mark:] " +
-					"Remove; NFC");
-
-	}
+	private final Transliterator _transliterator = Transliterator.getInstance(
+		"Greek-Latin; Cyrillic-Latin; NFD; [:Nonspacing Mark:] Remove; NFC");
 
 }
