+++ modules/apps/normalizer/normalizer-impl/src/test/java/com/liferay/normalizer/internal/NormalizerTest.java
@@ -12,8 +12,9 @@
  * details.
  */
 
-package com.liferay.util;
+package com.liferay.normalizer.internal;
 
+import com.liferay.normalizer.Normalizer;
 import com.liferay.petra.string.StringBundler;
 
 import org.junit.Assert;
@@ -31,7 +32,7 @@ public class NormalizerTest {
 
 		Assert.assertEquals(
 			"ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz",
-			Normalizer.normalizeToAscii(asciiAlphabet));
+			_normalizer.normalizeToAscii(asciiAlphabet));
 	}
 
 	@Test
@@ -46,7 +47,7 @@ public class NormalizerTest {
 
 		Assert.assertEquals(
 			"ABGDEZETHIKLMN'XOPRSTYPHCHPSOabgdezethiklmn'xoprsstyphchpso",
-			Normalizer.normalizeToAscii(greekAlphabet));
+			_normalizer.normalizeToAscii(greekAlphabet));
 	}
 
 	@Test
@@ -59,7 +60,7 @@ public class NormalizerTest {
 		Assert.assertEquals(
 			"AABCDEEFGHIIJKLMNOOOOPQRSTUUUUVWXYZaabcdeefghiijklmnoooopqrstuuu" +
 				"uvwxyz",
-			Normalizer.normalizeToAscii(hungarianAlphabet));
+			_normalizer.normalizeToAscii(hungarianAlphabet));
 	}
 
 	@Test
@@ -77,7 +78,7 @@ public class NormalizerTest {
 		Assert.assertEquals(
 			"ABVGDEEZZIJKLMNOPRSTUFHCCSS\"Y'EUAabvgdeezzijklmnoprstufhccss\"y" +
 				"'eua",
-			Normalizer.normalizeToAscii(russianAlphabet));
+			_normalizer.normalizeToAscii(russianAlphabet));
 	}
 
 	@Test
@@ -87,7 +88,9 @@ public class NormalizerTest {
 
 		Assert.assertEquals(
 			"ABCDEFGIJKLMNNOPQRSTUVWXYZabcdefghijklmnnopqrstuvwxyz",
-			Normalizer.normalizeToAscii(spanishAlphabet));
+			_normalizer.normalizeToAscii(spanishAlphabet));
 	}
+
+	private final Normalizer _normalizer = new NormalizerImpl();
 
 }
